Move Plugin class to be part of config loading, independent of File.
