+++ packages/babel-core/src/transformation/file/options/option-manager.js
@@ -1,5 +1,5 @@
 import * as context from "../../../index";
-import Plugin from "../../plugin";
+import Plugin from "./plugin";
 import * as messages from "babel-messages";
 import resolvePlugin from "../../../helpers/resolve-plugin";
 import resolvePreset from "../../../helpers/resolve-preset";
